We try to figure it out the proper values of different parameters
The best value of dt (which leads to stable evolution of the grap and also will not be so small - to speed up simulation). It turns out that a good choce is dt in range [0.001, 0.005].

Also we try to figure it out the range of possible values of Deff = D/beta. So it seems a good chice is a sth in range [0.5, 50]. This should give us the all interesting casses that we need.